pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "maskcloud"
version = "0.1.1"
description = "Just-In-Time Privacy Middleware for AI Agents. Format-preserving encryption with pluggable vault backends."
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"

authors = [
  { name = "Mask AI Solutions" }
]

keywords = ["privacy", "llm", "ai-agents", "fpe", "pii", "encryption", "soc2", "hipaa"]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Security :: Cryptography",
    "Programming Language :: Python :: 3",
]

dependencies = [
    "cryptography>=46.0.5",
    "presidio-analyzer>=2.2.361",
    "presidio-anonymizer>=2.2.361",
    "pydantic>=2.12.5",
]

[project.optional-dependencies]
redis = ["redis>=5.0"]
dynamodb = ["boto3>=1.34"]
memcached = ["pymemcache>=4.0"]
adk = ["google-adk>=1.0"]
langchain = ["langchain-core>=0.2"]
llamaindex = ["llama-index-core>=0.10"]
spacy = ["spacy>=3.4.4"]
sm = ["spacy>=3.4.4"]
md = ["spacy>=3.4.4"]
lg = ["spacy>=3.4.4"]
all = [
    "redis>=5.0",
    "boto3>=1.34",
    "pymemcache>=4.0",
    "google-adk>=1.0",
    "langchain-core>=0.2",
    "llama-index-core>=0.10",
    "spacy>=3.4.4",
]
examples = [
    "redis>=5.0.0",
    "litellm>=1.74",
    "python-dotenv>=1.0",
    "uvicorn>=0.30",
    "fastapi>=0.111",
    "httpx>=0.27",
]

[tool.hatch.build.targets.wheel]
packages = ["mask"]

[tool.hatch.metadata]
# Standard metadata only

[dependency-groups]
dev = [
    "langchain-core>=1.2.16",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "fakeredis>=2.21",
    "boto3>=1.42.59",
    "llama-index-core>=0.14.15",
    "pymemcache>=4.0.0",
    "spacy>=3.4.4",
]


[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "strict"

